requirements.txt
albumentations==2.0.5
beautifulsoup4==4.13.4
comet_ml==3.49.8
coremltools==8.2
dominate==2.9.1
Flask==3.1.0
GitPython==3.1.43
GitPython==3.1.44
h5py==3.13.0
image_slicer==2.1.1
imageio==2.31.2
imgaug==0.4.0
ipython==7.34.0
keras==3.9.2
lmdb==1.5.1
matplotlib==3.5.3
mss==10.0.0
nncf==2.16.0
numpy==1.21.5
nystrom_attention==0.0.11
onnx==1.17.0
onnxruntime==1.21.1
onnxsim==0.4.36
opencv_python==4.10.0.84
openvino==2025.1.0
paddle==1.0.2
pafy==0.5.5
pandas==1.3.5
Pillow==9.4.0
Pillow==11.2.1
pip==22.3.1
psutil==6.0.0
pycocotools==2.0.8
pytorch_fid==0.3.0
pyvips==2.2.3
PyYAML==6.0
PyYAML==6.0.2
Requests==2.32.3
scikit_image==0.19.3
scikit_learn==1.6.1
scipy==1.7.3
seaborn==0.13.2
setuptools==65.6.3
six==1.16.0
spams==2.6.5.4
staintools==2.1.2
tensorflow==2.19.0
tensorflowjs==4.22.0
tensorrt==10.9.0.34
tflite_runtime==2.14.0
tflite_support==0.4.4
thop==0.1.1.post2209072238
tifffile==2021.11.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.4
tritonclient==2.56.0
ultralytics==8.0.145
visdom==0.2.4
x2paddle==1.6.0
